Add : 샘플단어장 라우터 추가
일단 모델로 생성완료, 차근차근 3계층으로 만들겠습니다
